Change folder structure to locate files related to websocket on */websocket and related to http on */http
